analytics_service/requirements.txt
anyio==3.6.1
aiohttp~=3.8.1
asgiref==3.5.2
cassandra-driver==3.25.0
click==8.1.3
fastapi==0.78.0
geomet
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
pydantic==1.9.1
python-dotenv==0.20.0
six==1.16.0
sniffio==1.2.0
starlette
typing_extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
datapane==0.14.0
plotly==5.8.0
pandas
PyCryptodome
